Solved issue #6

+++ app/src/main/java/com/rohitthebest/manageyourrenters/ui/fragments/borrower/BorrowerHomeFragment.kt
@@ -185,7 +185,6 @@ class BorrowerHomeFragment : Fragment(R.layout.fragment_borrower_home),
                 if (!borrowerForMenus.isSynced) {
 
                     borrowerViewModel.deleteBorrower(
-                        requireContext(),
                         borrowerForMenus
                     )
                 } else {
@@ -193,7 +192,6 @@ class BorrowerHomeFragment : Fragment(R.layout.fragment_borrower_home),
                     if (isInternetAvailable(requireContext())) {
 
                         borrowerViewModel.deleteBorrower(
-                            requireContext(),
                             borrowerForMenus
                         )
                     } else {
@@ -278,7 +276,6 @@ class BorrowerHomeFragment : Fragment(R.layout.fragment_borrower_home),
 
                 // call the viewmodel method for adding or replacing the document
                 borrowerViewModel.addOrReplaceBorrowerSupportingDocument(
-                    requireContext(),
                     borrowerForMenus,
                     supportingDocumentHelperModel
                 )
@@ -296,28 +293,36 @@ class BorrowerHomeFragment : Fragment(R.layout.fragment_borrower_home),
             && borrowerForMenus.supportingDocument != null
         ) {
 
-            if (borrowerForMenus.supportingDocument?.documentType != DocumentType.URL) {
+            showAlertDialogForDeletion(
+                requireContext(),
+                {
+                    if (borrowerForMenus.supportingDocument?.documentType != DocumentType.URL) {
 
-                if (!isInternetAvailable(requireContext())) {
+                        if (!isInternetAvailable(requireContext())) {
 
-                    showToast(
-                        requireContext(),
-                        "Network connection required to delete the document from cloud"
-                    )
-                    return
-                } else {
+                            showToast(
+                                requireContext(),
+                                getString(R.string.network_required_for_deleting_file_from_cloud)
+                            )
+                            return@showAlertDialogForDeletion
+                        } else {
 
-                    deleteFileFromFirebaseStorage(
-                        requireContext(),
-                        borrowerForMenus.supportingDocument?.documentUrl!!
-                    )
+                            deleteFileFromFirebaseStorage(
+                                requireContext(),
+                                borrowerForMenus.supportingDocument?.documentUrl!!
+                            )
+                        }
+                    }
+                    borrowerForMenus.supportingDocument = null
+                    borrowerForMenus.isSupportingDocAdded = false
+                    borrowerViewModel.updateBorrower(borrowerForMenus)
+                    showToast(requireContext(), getString(R.string.supporting_document_deleted))
+                    it.dismiss()
+                },
+                {
+                    it.dismiss()
                 }
-            }
-            borrowerForMenus.supportingDocument = null
-            borrowerForMenus.isSupportingDocAdded = false
-            borrowerViewModel.updateBorrower(requireContext(), borrowerForMenus)
-            showToast(requireContext(), "Supporting Document deleted")
-
+            )
         } else {
             showToast(requireContext(), getString(R.string.no_supporting_doc_added))
         }
@@ -341,7 +346,7 @@ class BorrowerHomeFragment : Fragment(R.layout.fragment_borrower_home),
                         borrowerForMenus.key
                     )
 
-                    borrowerViewModel.updateBorrower(requireContext(), borrowerForMenus)
+                    borrowerViewModel.updateBorrower(borrowerForMenus)
 
                     borrowerAdapter.notifyItemChanged(currentAdapterPosition)
                 }
